// src/app/core/models/profile.model.ts
export interface PerfilMaestro {
  id?: number;
  usuario_id?: number;
  nombre?: string;
  apellido?: string;
  telefono?: string;
  curp?: string;
  rfc?: string;
  especialidad?: string;
  carrera?: string;
  direccion?: string;
  codigo_classroom?: string;
  foto_url?: string;
  nombre_usuario?: string;
  email?: string;
  existe?: boolean;
}